% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_3.R
\name{plus_three}
\alias{plus_three}
\title{Adds three to any number}
\usage{
plus_three(x)
}
\arguments{
\item{x}{an integer or double; any number}
}
\value{
a number
}
\description{
Adds three to any number
}
\examples{
plus_two(1)

}
