% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plus_two}
\alias{plus_two}
\title{Adds two to any number}
\usage{
plus_two(x)
}
\arguments{
\item{x}{an integer or double; any number}
}
\value{
a number
}
\description{
Adds two to any number
}
\examples{
plus_two(1)

}
